requirements.txt
# Основные зависимости
fastapi==0.115.0
uvicorn==0.34.2
psycopg2-binary==2.9.10
python-dotenv==1.1.0
pydantic==1.10.13
httpx==0.28.1

# Дополнительные зависимости для разработки
pytest==8.3.5
pylint==3.3.7 